requirements.dev.txt
django-debug-toolbar~=2.2
ipython~=7.13.0
